# src/config/swagger.yaml
openapi: 3.0.3
info:
  title: i-Station API
  description: i-Station 팀 협업 플랫폼 API 명세서
  version: 1.0.0
servers:
  - url: /
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: 회원가입
      description: 이메일, 비밀번호, 이름, 전화번호, 깃허브를 받아 유저를 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name, phone]
              properties:
                email:
                  type: string
                  format: email
                  example: mail@istation.dev
                password:
                  type: string
                  minLength: 8
                  example: password123!
                name:
                  type: string
                  minLength: 2
                  maxLength: 20
                  example: 홍길동
                phone:
                  type: string
                  example: 010-1234-5678
                github_url:
                  type: string
                  nullable: true
                  example: https://github.com/username
      responses:
        "201":
          description: 회원가입 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserResponse"
                      meta:
                        nullable: true
                        example: null
        "409":
          description: 이미 사용 중인 이메일
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      tags: [Auth]
      summary: 로그인
      description: JWT 토큰 발급 및 유저 정보(team_id 포함) 반환. 로그인 시 상태가 활동으로 변경됩니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: mail@istation.dev
                password:
                  type: string
                  example: password123!
      responses:
        "200":
          description: 로그인 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
                            example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                          user:
                            $ref: "#/components/schemas/UserResponse"
                      meta:
                        nullable: true
                        example: null
        "401":
          description: 이메일 혹은 비밀번호가 일치하지 않습니다
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      tags: [Auth]
      summary: 로그아웃
      description: 토큰 무효화. 로그아웃 시 상태가 자리비움으로 변경됩니다.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: 로그아웃 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/google:
    post:
      tags: [Auth]
      summary: 구글 OAuth 로그인 / 회원가입
      description: |
        구글 OAuth 인증을 처리합니다.
        - 최초 로그인 시: 구글에서 제공하는 Email, Name, Profile Image 데이터를 받아 DB에 저장 후 팀 로비로 이동합니다.
        - 기존 가입자: 구글 계정 고유 ID(google_uid)를 대조하여 일치하는 계정이 있으면 즉시 로그인 처리합니다.
        - 자체 이메일로 이미 가입된 계정과 이메일이 같을 경우 기존 계정에 연동하거나 안내 메시지를 반환합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_token]
              properties:
                id_token:
                  type: string
                  description: 구글 OAuth ID 토큰
                  example: ya29.a0AfH6SMC...
      responses:
        "200":
          description: 구글 로그인 / 회원가입 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
                            example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                          user:
                            $ref: "#/components/schemas/UserResponse"
                      meta:
                        nullable: true
                        example: null
        "409":
          description: 이미 자체 이메일로 가입된 계정
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ──────────────────────────────────────────
  # 2. Users
  # ──────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: 인증 토큰이 없거나 유효하지 않습니다
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            data: null
            meta: null
            error: 인증 토큰이 없거나 유효하지 않습니다
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          description: 실제 응답 데이터 (단일 객체 또는 배열)
        meta:
          nullable: true
          description: 페이지네이션 등 메타 정보 (없을 경우 null)
          type: object
          properties:
            total_count:
              type: integer
              example: 100
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
            total_pages:
              type: integer
              example: 10
        error:
          nullable: true
          example: null
    MessageResponse:
      type: object
      description: 메시지만 반환하는 성공 응답
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            message:
              type: string
              example: 성공적으로 처리되었습니다.
        meta:
          nullable: true
          example: null
        error:
          nullable: true
          example: null
    ErrorResponse:
      type: object
      description: 에러 응답
      properties:
        success:
          type: boolean
          example: false
        data:
          nullable: true
          example: null
        meta:
          nullable: true
          example: null
        error:
          type: string
          example: 오류가 발생했습니다.

    # ── 유저 ──
    UserResponse:
      type: object
      properties:
        uuid:
          type: integer
          example: 1
        name:
          type: string
          example: 홍길동
        profile_image:
          type: string
          nullable: true
          example: https://s3.amazonaws.com/profile.jpg

    # ── 팀 ──
...